Added another pure python alternative
Hi,
we (abusix) have recently decided to open source our pure python implementation of the Aho-Corasick algorithm. 
Its README also includes some performance comparisons and while your library is definitely the goto library, ahocorapy might also be useful for some people that need to stay clear of C-Extensions, are using pypy or need unicode support while using python2.7

Thanks for taking a look and it'd be nice to be included in your README, as it seems like a good place to find out a lot about Aho-Corasick implementations in python.