!feat: use glob-style wildcard patterns
BREAKING CHANGE: This commit changes the signature and behavior
of the Automaton keys, values, and items methods.

This commit changes the signatures of Automaton keys, values, and
items methods to the following:

- keys([pattern, [how]])
- values([pattern, [how]])
- items([pattern, [houw]])

This commit changes the behavior of these APIs to process patterns
as Unix shell-style glob patterns. The optional 'pattern' parameter
can contain one or more of the following wildcards anywhere in
the pattern:

- '?': matches any single character.
- '*': matches everything.

The '\\' character is reserved as the escape character. For example,
r'\\*' matches the literal '*' character. Each literal backslash must
be either expressed as '\\\\' or as a raw r'\\' string. For example,
r'\\t' matches '\\t'.

The optional 'how' argument is used to control how patterns are
matched using one of the following values:

- 'ahocorasick.MATCH_PREFIX' (default): Yield keys with prefixes that
  match the given pattern.
- 'ahocorasick.MATCH_WHOLE': Yield keys that fully match the given
  pattern.