Fix incorrect types passed to Py_BuildValue
In all cases, these calls told Python it was getting ints (`i`), which
are 4 bytes, but several cases passed either `Py_ssize_t` or
`Py_uintptr_t`, which are 8 bytes.

On little-endian machines, this will work, as the bytes align, but on
big-endian machine, Python will most probably see 0, since the top bytes
are more likely to be unfilled unless values are very large.